Ricerca rotta :-1:  | Aggiuyo controllo alfabetico nei test

+++ ArticoliWebService/Dtos/ArticoliDTO.cs
@@ -9,7 +9,7 @@ namespace ArticoliWebService.Dtos
         public string Descrizione { get; set; }
         public string Um { get; set; }
         public string CodStat { get; set; }
-        public Int16? PzCart { get; set; }
+        public Int32? PzCart { get; set; }
         public double? PesoNetto { get; set; }
         public DateTime? DataCreazione { get; set; }
